Add files via upload
This is an updated application. It has an additional app route and logic for it to go with the dashboard.html that is for the / route which shows a list of active tickets. There is a filter included for a specific user in our instance so you probably wont need that in your app if you choose to use this